+++ src/AuditableDomainEntity/Events/EntityEvents/AuditableEntityUpdated.cs
@@ -11,7 +11,7 @@ public record AuditableEntityUpdated(
     Ulid EntityId,
     Ulid? FieldId,
     Ulid? ParentId,
-    int EventVersion,
+    float EventVersion,
     List<IDomainValueFieldEvent>? ValueFieldEvents,
     List<IDomainEntityFieldEvent>? EntityFieldEvents,
     DateTimeOffset CreatedAtUtc)
